\name{RMS}
\alias{RMS}
\title{
Calculate the Root Mean Square of a Vector
}
\description{
Calculate the Root Mean Square of a Vector after removing NAs from the
vector.
}
\usage{
RMS(num)
}
\arguments{
  \item{num}{
Numeric Vector
}
}
\value{
Returns the root mean square of the numeric vector
}
\author{
Taran Esplin
}
\examples{
RMS(1:10)
sd(1:10)
}
\keyword{ univar }% __ONLY ONE__ keyword per line
